requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
appdirs==1.4.4            # via black
appnope==0.1.0            # via ipython
attrs==19.3.0             # via black, pytest
backcall==0.1.0           # via ipython
black==19.10b0            # via -r requirements/dev.in
click==7.1.2              # via black, pip-tools
coverage==5.1             # via pytest-cov
decorator==4.4.2          # via ipython, traitlets
factory-boy==2.12.0       # via pytest-factoryboy
faker==4.1.0              # via factory-boy
flake8==3.8.2             # via -r requirements/dev.in
freezegun==0.3.15         # via pytest-freezegun
inflection==0.4.0         # via pytest-factoryboy
ipython-genutils==0.2.0   # via traitlets
ipython==7.15.0           # via -r requirements/dev.in
isort==4.3.21             # via -r requirements/dev.in
jedi==0.17.0              # via ipython
mccabe==0.6.1             # via flake8
more-itertools==8.3.0     # via pytest
packaging==20.4           # via pytest
parso==0.7.0              # via jedi
pathspec==0.8.0           # via black
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.2.0          # via -r requirements/dev.in
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.5     # via ipython
ptyprocess==0.6.0         # via pexpect
py==1.8.1                 # via pytest
pycodestyle==2.6.0        # via flake8
pyflakes==2.2.0           # via flake8
pygments==2.6.1           # via ipython
pyparsing==2.4.7          # via packaging
pytest-cov==2.9.0         # via -r requirements/dev.in
pytest-django==3.9.0      # via -r requirements/dev.in
pytest-factoryboy==2.0.3  # via -r requirements/dev.in
pytest-freezegun==0.4.2   # via -r requirements/dev.in
pytest==5.4.3             # via -r requirements/dev.in, pytest-cov, pytest-django, pytest-factoryboy, pytest-freezegun
python-dateutil==2.8.1    # via faker, freezegun
regex==2020.5.14          # via black
six==1.15.0               # via freezegun, packaging, pip-tools, python-dateutil, traitlets
text-unidecode==1.3       # via faker
toml==0.10.1              # via black
traitlets==4.3.3          # via ipython
typed-ast==1.4.1          # via black
wcwidth==0.2.3            # via prompt-toolkit, pytest

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
